Implemented flow control for printing.  This is switchable with the FLOW_CONTROL boolean, and is set to OFF (false).
If flow control is on, each frame in printing processes ONLY will be sequenced and validated.  The transmit process will only continue when a response has confirmed correct receipt of the previous frame.  The system will attempt to send each frame for a maximum of 5 attempts, and then abort. An abort will abort the whole print queue.

It is advisable to send an ERASE command at the start of each print queue, to guarantee that the correct bitmap will be printed.

This commit still contains debug log commands.